requirements.txt
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Script==2.0.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
PyYAML==3.12
six==1.10.0
Werkzeug==0.12.2
